backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
pymongo==4.6.1
marshmallow==3.20.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2 